+++ tests/unit/commands/test_argtypes.py
@@ -11,10 +11,10 @@ import pytest
 from vimiv.commands import argtypes
 
 
-def test_scroll_direction():
+@pytest.mark.parametrize("name", ("left", "right", "up", "down"))
+def test_scroll_direction(name):
     # Would raise exception if a name is invalid
-    # pylint: disable=expression-not-assigned
-    [argtypes.Direction(name) for name in ["left", "right", "up", "down"]]
+    assert isinstance(argtypes.Direction(name), argtypes.Direction)
 
 
 def test_fail_scroll_direction():
@@ -22,10 +22,10 @@ def test_fail_scroll_direction():
         argtypes.Direction("other")
 
 
-def test_zoom():
+@pytest.mark.parametrize("name", ("in", "out"))
+def test_zoom(name):
     # Would raise exception if a name is invalid
-    # pylint: disable=expression-not-assigned
-    [argtypes.Zoom(name) for name in ["in", "out"]]
+    assert isinstance(argtypes.Zoom(name), argtypes.Zoom)
 
 
 def test_fail_zoom():
@@ -33,22 +33,45 @@ def test_fail_zoom():
         argtypes.Zoom("other")
 
 
-def test_image_scale_text():
+@pytest.mark.parametrize("name", ("fit", "fit-width", "fit-height"))
+def test_image_scale_text(name):
     # Would raise exception if a name is invalid
-    # pylint: disable=expression-not-assigned
-    [argtypes.ImageScaleFloat(name) for name in ["fit", "fit-width", "fit-height"]]
+    assert isinstance(argtypes.ImageScaleFloat(name), argtypes.ImageScale)
 
 
 def test_image_scale_float():
     assert argtypes.ImageScaleFloat("0.5") == 0.5
 
 
-def test_command_history_direction():
+@pytest.mark.parametrize("name", ("next", "prev"))
+def test_command_history_direction(name):
     # Would raise exception if a name is invalid
-    # pylint: disable=expression-not-assigned
-    [argtypes.HistoryDirection(name) for name in ["next", "prev"]]
+    assert isinstance(argtypes.HistoryDirection(name), argtypes.HistoryDirection)
 
 
 def test_fail_command_history_direction():
     with pytest.raises(ValueError, match="not a valid HistoryDirection"):
         argtypes.HistoryDirection("other")
+
+
+@pytest.mark.parametrize("size", ((3, 3), (4, 3), (16, 9)))
+@pytest.mark.parametrize("separator", argtypes.AspectRatio.SEPARATORS)
+def test_aspectratio(size, separator):
+    definition = separator.join(str(length) for length in size)
+    width, height = size
+    aspectratio = argtypes.AspectRatio(definition)
+    assert aspectratio.width() == int(width)
+    assert aspectratio.height() == int(height)
+    assert not aspectratio.keep
+
+
+@pytest.mark.parametrize("definition", ("4to3", "4:3:2", "42", "hello:world"))
+def test_fail_aspectratio(definition):
+    with pytest.raises(ValueError, match="Invalid aspectratio"):
+        argtypes.AspectRatio(definition)
+
+
+@pytest.mark.parametrize("value", ("Keep", "keep", "keeP"))
+def test_aspectratio_keep(value):
+    aspectratio = argtypes.AspectRatio(value)
+    assert aspectratio.keep