+++ vimiv/imutils/imtransform.py
@@ -169,6 +169,9 @@ class Transform(QTransform):
         )
         self._apply(transformed.copy(rect))
 
+    def crop(self, rect):
+        self._apply(self.current.copy(rect))
+
     def _apply(self, transformed):
         """Check the transformed pixmap for validity and apply it to the handler."""
         if transformed.isNull():
@@ -185,7 +188,11 @@ class Transform(QTransform):
     @property
     def changed(self):
         """True if transformations have been applied."""
-        return not self.isIdentity()
+        transformed = not self.isIdentity()
+        if self._original is None:
+            return transformed
+        cropped = self.current.rect() != self._original.rect()
+        return transformed or cropped
 
     @property
     def matrix(self):