+++ docs/changelog.rst
@@ -266,6 +266,11 @@ Added:
 * A new ``PromptSetting`` type which is essentially a boolean setting with the
   additional ``ask`` value. If the value is ``ask``, the user is prompted everytime the
   boolean state of this setting is requested.
+* The ``:crop`` command which displays a rectangle to crop the curent image. The
+  rectangle can be dragged and resized using the mouse. As with ``:straighten``, accept
+  the changes with ``<return>`` and reject them with ``<escape>``. The
+  ``{transformation-info}`` status module displays the currently selected geometry of
+  the original image.
 
 Changed:
 ^^^^^^^^